refactor/챌린지 생성 fragment로 변경 및 UI수정
