+++ app/src/main/AndroidManifest.xml
@@ -51,21 +51,25 @@
 
         <activity
             android:name=".view.activities.CreateChallengeActivity"
-            android:exported="true" />
+            android:exported="true" >
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
 
-        <activity
-            android:name=".view.activities.CreateCrew1Activity"
-            android:exported="true" />
+        <activity android:name=".view.activities.CreateCrewActivity"
+            android:exported="true">
 
-        <activity
-            android:name=".view.activities.CreateCrew2Activity"
-            android:exported="true" />
+        </activity>
 
         <activity android:name=".view.activities.SignUpActivity"/>
 
         <activity
             android:name=".view.activities.CreateSolo1Activity"
-            android:exported="true" />
+            android:exported="true" >
+
+        </activity>
 
         <activity
             android:name=".view.activities.CreateSolo2Activity"
@@ -74,10 +78,6 @@
         <activity
             android:name=".view.activities.LoginActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
 
         </activity>
