+++ app/src/main/java/com/example/yourun/view/activities/CalendarActivity.kt
@@ -16,8 +16,6 @@ import androidx.lifecycle.lifecycleScope
 import com.example.yourun.R
 import com.example.yourun.model.data.RunningStatsResponse
 import com.example.yourun.model.network.ApiClient
-import com.example.yourun.model.network.ApiResponse
-import com.google.android.gms.common.api.Response
 import kotlinx.coroutines.launch
 import java.time.LocalDate
 import java.time.YearMonth
@@ -71,7 +69,7 @@ class CalendarActivity : AppCompatActivity() {
         fetchRunningStats(selectedDate.year, selectedDate.monthValue)
 
         // View 초기화
-        weekCalendarGrid = findViewById(R.id.weekCalendarGrid)
+        weekCalendarGrid = findViewById(R.id.dateLayout)
         fullCalendarGrid = findViewById(R.id.fullCalendarGrid)
         dragBar = findViewById(R.id.dragBar)
