+++ app/src/main/res/layout/activity_create_solo1.xml
@@ -200,7 +200,7 @@
 
             <!-- 주간 달력 그리드 -->
             <GridLayout
-                android:id="@+id/weekCalendarGrid"
+                android:id="@+id/dateLayout"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:columnCount="7"
