+++ app/src/main/res/layout/fragment_create_crew1.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools">
 
-
+    <!-- Top Bar -->
     <FrameLayout
         android:id="@+id/top_bar"
         android:layout_width="wrap_content"
@@ -46,11 +46,9 @@
             android:layout_marginEnd="20dp"
             android:background="@drawable/ic_calendar" />
 
-
-
     </FrameLayout>
 
-
+    <!-- Crew Challenge Text -->
     <TextView
         android:id="@+id/tv_crew1"
         android:layout_width="wrap_content"
@@ -65,6 +63,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/top_bar" />
 
+    <!-- Other Crew Challenge Texts -->
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -84,7 +83,6 @@
         app:layout_constraintStart_toStartOf="@id/tv_crew1"
         app:layout_constraintTop_toTopOf="@id/tv_crew1" />
 
-
     <TextView
         android:id="@+id/tv_crew2"
         android:layout_width="wrap_content"
@@ -96,6 +94,7 @@
         app:layout_constraintStart_toStartOf="@id/tv_crew1"
         app:layout_constraintTop_toTopOf="@id/tv_crew1" />
 
+    <!-- Edit Text Fields -->
     <EditText
         android:id="@+id/et_crew_name"
         android:layout_width="340dp"
@@ -108,7 +107,7 @@
         app:layout_constraintStart_toStartOf="@id/tv_crew2"
         app:layout_constraintTop_toBottomOf="@id/tv_crew2" />
 
-
+    <!-- Other Text and EditText fields -->
     <TextView
         android:id="@+id/textView5"
         android:layout_width="wrap_content"
@@ -119,7 +118,6 @@
         android:textStyle="bold"
         app:layout_constraintStart_toStartOf="@id/et_crew_name"
         app:layout_constraintTop_toBottomOf="@+id/et_crew_name" />
-
 
     <TextView
         android:id="@+id/tv_crew3"
@@ -143,7 +141,7 @@
         app:layout_constraintStart_toStartOf="@id/tv_crew3"
         app:layout_constraintTop_toBottomOf="@id/tv_crew3" />
 
-
+    <!-- Date Text -->
     <TextView
         android:id="@+id/tv_date"
         android:layout_width="wrap_content"
@@ -155,10 +153,10 @@
         app:layout_constraintStart_toStartOf="@id/tv_crew2"
         app:layout_constraintTop_toBottomOf="@+id/et_crew" />
 
-
+    <!-- Calendar Layout -->
     <LinearLayout
         android:id="@+id/calender"
-        android:layout_width="match_parent"
+        android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:orientation="vertical"
         android:padding="4dp"
@@ -168,7 +166,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tv_date">
 
-        <!-- 상단 년/월 표시 리니어레이아웃 -->
+        <!-- Month/Year Layout -->
         <LinearLayout
             android:id="@+id/monthYearLayout"
             android:layout_width="match_parent"
@@ -176,7 +174,6 @@
             android:orientation="horizontal"
             android:padding="8dp">
 
-            <!-- 현재 년/월 표시 -->
             <TextView
                 android:id="@+id/monthYearText"
                 android:layout_width="wrap_content"
@@ -188,65 +185,59 @@
                 android:textStyle="bold" />
         </LinearLayout>
 
-        <!-- 요일 표시 (날짜 앞에 나오는 일, 월, 화 등) -->
+        <!-- Weekday Layout -->
         <LinearLayout
             android:id="@+id/weekdayLayout"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:gravity="center"
             android:orientation="horizontal"
-            android:padding="8dp"
+            android:padding="5dp"
             android:useDefaultMargins="false">
-            <!-- 요일이 들어갈 TextView들을 동적으로 추가할 예정 -->
+
+            <!-- 요일을 동적으로 추가 -->
         </LinearLayout>
 
-        <!-- 주간 달력 -->
+        <!-- Calendar Grid -->
         <LinearLayout
             android:id="@+id/dragBarContainer"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="0dp"
             android:gravity="center"
             android:orientation="vertical"
             android:padding="0dp"
-            android:useDefaultMargins="false">
+            android:useDefaultMargins="false"
+            android:clickable="true">
 
-            <!-- 주간 달력 그리드 -->
-            <GridLayout
-                android:id="@+id/weekCalendarGrid"
+            <LinearLayout
+                android:id="@+id/dateLayout"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:columnCount="7"
                 android:orientation="horizontal"
-                android:padding="8dp"
-                android:useDefaultMargins="false"></GridLayout>
+                android:padding="5dp"
+                android:useDefaultMargins="false"
+                android:clickable="true">
+
+                <!-- 날짜들이 동적으로 추가 -->
+            </LinearLayout>
 
         </LinearLayout>
     </LinearLayout>
+
+    <!-- Next Button -->
+
+    <!-- Bottom Navigation -->
 
     <ImageButton
         android:id="@+id/next_btn"
         android:layout_width="360dp"
         android:layout_height="48dp"
         android:layout_gravity="center"
-        android:layout_marginTop="235dp"
+        android:layout_marginTop="268dp"
         android:background="@drawable/btn_create"
         app:layout_constraintStart_toStartOf="@id/tv_date"
         app:layout_constraintTop_toTopOf="@id/tv_date" />
 
-
-
-
     <include
         layout="@layout/bottom_navigation" />
-
-
-
-
-
-
-
-
-
 </androidx.constraintlayout.widget.ConstraintLayout>
